Resolved compile warnings.

+++ src/test/java/net/sf/katta/client/MapFileClientTest.java
@@ -157,7 +157,7 @@ public class MapFileClientTest extends AbstractKattaTest {
       t.join();
     }
     long time = System.currentTimeMillis() - startTime;
-    System.out.println((1000.0 * (double) count.intValue() / (double) time) + " requests / sec");
+    System.out.println((1000.0 * count.intValue() / time) + " requests / sec");
     assertTrue(exceptions.isEmpty());
   }
